+++ packages/adp-tooling/test/unit/prompts/add-new-model/index.test.ts
@@ -156,7 +156,7 @@ describe('getPrompts', () => {
 
         expect(typeof validation).toBe('function');
         expect(validation?.('customer.testName')).toBe(
-            'An OData annotation or service with the same name was already added to the project. Rename and try again.'
+            'A data source and SAPUI5 model with the same name were already added to the project. Rename and try again.'
         );
     });