+++ packages/project-access/src/index.ts
@@ -6,7 +6,7 @@ export {
     MinCdsVersion,
     fioriToolsDirectory
 } from './constants';
-export { getFilePaths } from './file';
+export { getFilePaths, findFilesByExtension } from './file';
 export { normalizePath } from './path';
 export {
     addPackageDevDependency,
@@ -68,10 +68,11 @@ export {
     readFlexChanges,
     processServices,
     getMainService,
-    getGlobalCdsHomePath
+    getGlobalCdsHomePath,
+    isUI5IdUnique
 } from './project';
 export { execNpmCommand } from './command/npm-command';
 export * from './types';
-export * from './library';
+export { checkDependencies, getReuseLibs } from './library';
 export { findRecursiveHierarchyKey, getTableCapabilitiesByEntitySet } from './odata';
 export { hasDependency } from './project';