+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.9.45
+
+### Patch Changes
+
+- Updated dependencies [1d60871]
+    - @sap-ux/project-access@1.36.0
+    - @sap-ux/cds-odata-annotation-converter@0.7.16
+
+## 0.9.44
+
+### Patch Changes
+
+- 52f6549: Add .cds annotations support to enable linting of CAP apps with the eslint-plugin-fiori-tools.
+
 ## 0.9.43
 
 ### Patch Changes